rest_api/views/blog.py
from rest_framework import generics
from rest_framework.serializers import Serializer
from main_app.models import ContactForm
from rest_api.serializers import ContactSerializer
from rest_framework.response import Response
from rest_framework import status
from core.settings import env

