requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
